создал для комментариев отдельный компонент
