+++ src/components/film-detail.js
@@ -1,31 +1,7 @@
 import AbstractSmartComponent from "./abstract-smart-component.js";
-// import {generateFilmDetail} from "../mock/film.js";
-import {COMMENT_EMOJI} from "../mock/const";
 
-
-const createCommentsTemplate = (author, text, emoji, dayCommented, commentId) => {
-  return (
-    `<li class="film-details__comment">
-    <span class="film-details__comment-emoji">
-      <img src="./images/emoji/${COMMENT_EMOJI[emoji]}" width="55" height="55" alt="emoji-smile">
-    </span>
-    <div>
-      <p class="film-details__comment-text"> ${text}</p>
-      <p class="film-details__comment-info">
-        <span class="film-details__comment-author">${author}</span>
-        <span class="film-details__comment-day">${dayCommented}</span>
-        <button class="film-details__comment-delete" data-comment-id="${commentId}">Delete</button>
-      </p>
-    </div>
-  </li>`
-  );
-};
-
-
-const filmDetailTemplate = (film, emoji, comments) => {
+const filmDetailTemplate = (film) => {
   const {title, writers, rating, alternativeTitle, runtime, releaseCountry, actors, ageRating, dataRelease, director, genre, poster, description, isWatchList, isWatched, isFavorite} = film;
-  const emojiMarkup = emoji ? `<img src ="./images/emoji/${emoji}" alt="" width="55" height="55">` : ``;
-  const commentsMarkup = comments.map((comment) => createCommentsTemplate(comment.author, comment.comment, comment.emotion, comment.date, comment.id)).join(`\n`);
   return (
     `<section class="film-details">
     <form class="film-details__inner" action="" method="get">
@@ -104,42 +80,7 @@ const filmDetailTemplate = (film, emoji, comments) => {
       </div>
   
       <div class="form-details__bottom-container">
-        <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${film.comments.length}</span></h3>
-        <ul class="film-details__comments-list">
-          ${commentsMarkup}
-
-          <div class="film-details__new-comment">
-    <div for="add-emoji" class="film-details__add-emoji-label">
-    ${emojiMarkup}
-    <input type="hidden" name="add-emoji" value="">
-    </div>
-
-    <label class="film-details__comment-label">
-      <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-    </label>
-
-    <div class="film-details__emoji-list">
-      <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
-      <label class="film-details__emoji-label" for="emoji-smile">
-        <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
-      </label>
-
-      <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-      <label class="film-details__emoji-label" for="emoji-sleeping">
-        <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
-      </label>
-
-      <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-      <label class="film-details__emoji-label" for="emoji-puke">
-        <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
-      </label>
-
-      <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-      <label class="film-details__emoji-label" for="emoji-angry">
-        <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
-      </label>
-    </div>
+          
           </ul>
           </div>
         </section>
@@ -151,20 +92,13 @@ const filmDetailTemplate = (film, emoji, comments) => {
 };
 
 export default class FilmDetail extends AbstractSmartComponent {
-  constructor(film, comments) {
+  constructor(film) {
     super();
     this._film = film;
-    this._comments = comments;
     this._closeButtonHandler = null;
     this._setWatchListDetailHandler = null;
     this._setWatchedDetailHandler = null;
     this._setFavoriteDetailHandler = null;
-    this._deleteCommentButtonHandler = null;
-    this._setCommentHandler = null;
-    this._getNewComment = null;
-    this._emoji = null;
-
-    this._subscribeOnEvents();
 
   }
 
@@ -173,13 +107,10 @@ export default class FilmDetail extends AbstractSmartComponent {
     this.setWatchListButtonClickHandlerDetail(this._setWatchListDetailHandler);
     this.setWatchedButtonClickHandlerDetail(this._setWatchedDetailHandler);
     this.setFavoriteButtonClickHandlerDetail(this._setFavoriteDetailHandler);
-    this.deleteCommentButtonHandler(this._deleteCommentButtonHandler);
-    this.setCommentHandler(this._setCommentHandler);
-    this._subscribeOnEvents();
   }
 
   getTemplate() {
-    return filmDetailTemplate(this._film, this._emoji, this._comments);
+    return filmDetailTemplate(this._film);
   }
 
   onCloseButtonHandler(handler) {
@@ -207,44 +138,5 @@ export default class FilmDetail extends AbstractSmartComponent {
     .addEventListener(`click`, handler);
 
     this._setFavoriteDetailHandler = handler;
-  }
-
-  deleteCommentButtonHandler(handler) {
-    this.getElement().querySelectorAll(`.film-details__comment-delete`).forEach((it) => {
-      it.addEventListener(`click`, (event) => {
-        event.preventDefault();
-        const currentID = it.getAttribute(`data-comment-id`);
-        handler(+currentID);
-      });
-      this._deleteCommentButtonHandler = handler;
-    });
-  }
-
-  setCommentHandler(handler) {
-    this.getElement().querySelector(`.film-details__comment-input`)
-    .addEventListener(`keydown`, handler);
-    this._setCommentHandler = handler;
-  }
-
-  getNewComment(newComment) {
-    return {
-      author: `Some`,
-      text: newComment,
-      emoji: this._emoji,
-      dayCommented: `today`,
-      id: new Date().getSeconds() + Math.random(),
-    };
-  }
-
-
-  _subscribeOnEvents() {
-    const element = this.getElement();
-
-    element.querySelector(`.film-details__emoji-list`).addEventListener(`click`, (event) => {
-      if (event.target.value in COMMENT_EMOJI) {
-        this._emoji = COMMENT_EMOJI[event.target.value];
-        this.rerender();
-      }
-    });
   }
 }
