+++ src/controllers/movie-controller.js
@@ -1,5 +1,6 @@
 import CardComponent from "../components/film-card.js";
 import FilmDetailComponent from "../components/film-detail.js";
+import CommentsComponent from "../components/comments-list.js";
 
 
 import {render, replace, remove, RenderPosition} from "../utils/render.js";
@@ -15,31 +16,36 @@ export default class MovieController {
     this._commentsModel = commentsModel;
     this._cardComponent = null;
     this._filmDetailComponent = null;
+    this._commentsComponent = null;
     this._filmComments = [];
     this._onEscKeyDown = this._onEscKeyDown.bind(this);
-    this._resetForm = this._resetForm.bind(this);
+    // this._resetForm = this._resetForm.bind(this);
   }
 
   _openPopup() {
     this._onViewChange();
+    this._commentsComponent = new CommentsComponent(this._commentsModel);
     bodyElement.appendChild(this._filmDetailComponent.getElement());
+    const element = this._filmDetailComponent.getElement().querySelector(`.form-details__bottom-container`);
+    element.appendChild(this._commentsComponent.getElement());
   }
 
   setDefaultView() {
     this._filmDetailComponent.getElement().remove();
+    // console.log(this._commentsComponent);
   }
 
   _onEscKeyDown(evt) {
     if (evt.key === `Escape` || evt.key === `Ecs`) {
-      this._resetForm();
+      // this._resetForm();
       this.setDefaultView();
     }
   }
 
-  _resetForm() {
-    this._filmDetailComponent.getElement().querySelector(`.film-details__inner`).reset();
-    this._filmDetailComponent.getElement().querySelector(`.film-details__add-emoji-label`).innerHTML = ``;
-  }
+  // _resetForm() {
+  //   this._filmDetailComponent.getElement().querySelector(`.film-details__inner`).reset();
+  //   this._filmDetailComponent.getElement().querySelector(`.film-details__add-emoji-label`).innerHTML = ``;
+  // }
 
   destroy() {
     remove(this._cardComponent);
@@ -55,14 +61,8 @@ export default class MovieController {
     const oldFilmComponent = this._cardComponent;
     const oldDetailComponent = this._filmDetailComponent;
 
-    film.comments.forEach((comment) => {
-      const currentComment = this.getCommentsFromModel(comment);
-      this._filmComments.push(currentComment);
-    });
-
-
     this._cardComponent = new CardComponent(film);
-    this._filmDetailComponent = new FilmDetailComponent(film, this._filmComments);
+    this._filmDetailComponent = new FilmDetailComponent(film);
 
     if (oldFilmComponent && oldDetailComponent) {
       replace(this._cardComponent, oldFilmComponent);
@@ -101,24 +101,24 @@ export default class MovieController {
       this._onDataChange(this, film, Object.assign({}, film, {isFavorite: !film.isFavorite}));
     });
 
-    this._filmDetailComponent.deleteCommentButtonHandler((commentId) => {
-      this._commentsModel.removeComment(commentId);
-      // return this._onDataChange(this, film, Object.assign({}, film, {}));
-    });
+    // this._filmDetailComponent.deleteCommentButtonHandler((commentId) => {
+    //   this._commentsModel.removeComment(commentId);
+    //   // return this._onDataChange(this, film, Object.assign({}, film, {}));
+    // });
 
-    this._filmDetailComponent.setCommentHandler((event) => {
-      if (event.keyCode === 13 && event.ctrlKey) {
-        const form = this._filmDetailComponent.getElement().querySelector(`.film-details__inner`);
-        const formData = new FormData(form);
-        const comment = formData.get(`comment`);
-        const newComment = this._filmDetailComponent.getNewComment(comment);
-        this._commentsModel.addComment(newComment);
-        this._onDataChange(this, film, Object.assign({}, film, {}));
-      }
-    });
+    // this._filmDetailComponent.setCommentHandler((event) => {
+    //   if (event.keyCode === 13 && event.ctrlKey) {
+    //     const form = this._filmDetailComponent.getElement().querySelector(`.film-details__inner`);
+    //     const formData = new FormData(form);
+    //     const comment = formData.get(`comment`);
+    //     const newComment = this._filmDetailComponent.getNewComment(comment);
+    //     this._commentsModel.addComment(newComment);
+    //     this._onDataChange(this, film, Object.assign({}, film, {}));
+    //   }
+    // });
 
     this._filmDetailComponent.onCloseButtonHandler(() => {
-      this._resetForm();
+      // this._resetForm();
       this.setDefaultView();
       document.removeEventListener(`keydown`, this._onEscKeyDown);
     });
